Initial Upload
Just drag the Quick Asset Import folder into your project, then go to Window > Quick Asset.
In order for this to work, assets must have been previously downloaded, it won't check for updates, it won't download updates. It just quickly gathers the localized version where Unity stores it and just 'Run's' the package file.

There may be some kinks in it here and there, such as when it's done, it clears the window out. Couldn't find a way around that.

TESTED ONLY ON WINDOWS